feat : transfer diamond and metad bettwen FE and BE (#1640)

+++ Assets/Scripts/Networking/Menu/DimensionBox/MetadResponseData.cs
@@ -28,7 +28,7 @@ namespace CryptoQuest.Networking.Menu.DimensionBox
         public int Gold { get; set; }
 
         [JsonProperty("diamond")]
-        public int Diamond { get; set; }
+        public float Diamond { get; set; }
 
         [JsonProperty("soul")]
         public int Soul { get; set; }
